+++ packages/ui/src/App.svelte
@@ -212,6 +212,7 @@
             }}
             subscribingText={t('Subscribing')}
             waitingText={t('Waiting')}
+            reconnectingText={t('Reconnecting')}
             allConnectedText={t('All interfaces are connected')}
             someNotConnectedText={t('Not every interface is connected')}
             portLabel={t('Port')}